Structurise project into application factory pattern - add app and helper func as modules, run from separate file
