+++ todo/helpers/hash/hash.py
